fix: fixing patch and delete departement

+++ api/v1/controllers/department_controller.go
@@ -3,6 +3,7 @@ package controllers
 import (
 	"database/sql"
 	"encoding/json"
+	"fmt"
 	"log"
 	"strconv"
 	"unicode/utf8"
@@ -10,8 +11,10 @@ import (
 	"net/http"
 
 	"github.com/gasBlar/GoGoManager/api/v1/services"
+	"github.com/gasBlar/GoGoManager/db"
 	"github.com/gasBlar/GoGoManager/models"
 	"github.com/gasBlar/GoGoManager/utils"
+	"github.com/go-playground/validator/v10"
 	"github.com/gorilla/mux"
 )
 
@@ -64,7 +67,10 @@ func (c *DepartmentController) CreateDepartment(w http.ResponseWriter, r *http.R
 
 func (c *DepartmentController) PatchDepartment(w http.ResponseWriter, r *http.Request) {
 	// user := r.Context().Value("user").(*utils.Claims)
-
+	if r.Header.Get("Content-Type") != "application/json" {
+		http.Error(w, "Invalid Content-Type, must be application/json", http.StatusBadRequest)
+		return
+	}
 	vars := mux.Vars(r)
 	departmentId := vars["departmentId"]
 
@@ -74,20 +80,49 @@ func (c *DepartmentController) PatchDepartment(w http.ResponseWriter, r *http.Re
 		return
 	}
 
+	if len(department.Name) < 4 {
+		utils.Response(w, http.StatusBadRequest, "", nil)
+		return
+	}
+
+	err := validateUpdateRequestDep(department)
+	if err != nil {
+		utils.Response(w, http.StatusBadRequest, fmt.Sprintf("Validation failed: %v", err), nil)
+		return
+	}
+
+	database := db.DB
+	var id string
+	res := database.QueryRow("SELECT id FROM department WHERE id = ?", departmentId).Scan(&id)
+	if res == sql.ErrNoRows {
+		utils.Response(w, http.StatusNotFound, "Id", nil)
+	}
+
 	if err := c.Service.PatchDepartment(departmentId, &department); err != nil {
 		log.Println("Error Updating department:", err)
 		http.Error(w, "Error updating department", http.StatusInternalServerError)
 		return
 	}
 
-	w.WriteHeader(http.StatusOK)
-	json.NewEncoder(w).Encode(map[string]string{"message": "Department updated successfully"})
+	response := map[string]string{
+		"departmentId": departmentId,
+		"name":         department.Name,
+	}
+	utils.Response(w, http.StatusOK, "", response)
 }
 
 func (c *DepartmentController) DeleteDepartment(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	departmentId := vars["departmentId"]
 	// user := r.Context().Value("user").(*utils.Claims)
+
+	var id string
+	var name string
+	res := db.DB.QueryRow("SELECT id, name FROM department WHERE id = ?", departmentId).Scan(&id, &name)
+	if res == sql.ErrNoRows {
+		utils.Response(w, http.StatusNotFound, "Id", nil)
+		return
+	}
 
 	if err := c.Service.DeleteDepartment(departmentId); err != nil {
 		if err.Error() == "access denied: manager does not have permission to modify this department" {
@@ -99,8 +134,11 @@ func (c *DepartmentController) DeleteDepartment(w http.ResponseWriter, r *http.R
 		return
 	}
 
-	w.WriteHeader(http.StatusOK)
-	// json.NewEncoder(w).Encode(map[string]string{"message": "Department deleted successfully"})
+	response := map[string]string{
+		"departmentId": id,
+		"name":         name,
+	}
+	utils.Response(w, http.StatusOK, "", response)
 }
 
 func GetDepartments(db *sql.DB) http.HandlerFunc {
@@ -130,4 +168,15 @@ func GetDepartments(db *sql.DB) http.HandlerFunc {
 			return
 		}
 	}
+}
+
+func validateUpdateRequestDep(req models.DepartmentPatch) error {
+	validate := validator.New()
+	validate.RegisterValidation("nonempty", validateNonEmpty)
+
+	err := validate.Struct(req)
+	if err != nil {
+		return err
+	}
+	return nil
 }
